moved fixtures to fixtures.py file. use rf fixture from pytest-django
