+++ somsolet/tests/conftest.py
@@ -1,10 +1,5 @@
 import psycopg2
-import pytest
-from django.db import connections
 from psycopg2.extensions import ISOLATION_LEVEL_AUTOCOMMIT
-
-from .factories import (CampaignFactory, EngineeringFactory,
-                        TechnicalDetailsFactory, UserFactory)
 
 
 def run_sql(sql):
@@ -13,23 +8,3 @@ def run_sql(sql):
     cur = conn.cursor()
     cur.execute(sql)
     conn.close()
-
-
-@pytest.fixture
-def engenieering_user(db):
-    return UserFactory()
-
-
-@pytest.fixture
-def engenieering(db):
-    return EngineeringFactory()
-
-
-@pytest.fixture
-def campaing__solar_paco(db):
-    return CampaignFactory()
-
-
-@pytest.fixture
-def technical_details(db):
-    return TechnicalDetailsFactory()
